+++ schematics/table/files/__path__/__name@dasherize@if-flat__/__name@dasherize__.component.html
@@ -20,7 +20,7 @@
   <mat-paginator #paginator
     [length]="dataSource.data.length"
     [pageIndex]="0"
-    [pageSize]="10"
+    [pageSize]="50"
     [pageSizeOptions]="[25, 50, 100, 250]">
   </mat-paginator>
 </div>
