+++ schematics/table/files/__path__/__name@dasherize@if-flat__/__name@dasherize__.component.ts
@@ -1,85 +1,7 @@
 import { Component, OnInit, ViewChild, <% if(!!viewEncapsulation) { %>, ViewEncapsulation<% }%><% if(changeDetection !== 'Default') { %>, ChangeDetectionStrategy<% }%> } from '@angular/core';
 import { Observable } from 'rxjs/Observable';
-import { map } from 'rxjs/operators/map';
 import { MatPaginator, MatSort } from '@angular/material';
-import { DataSource } from '@angular/cdk/collections';
-import { of } from 'rxjs/observable/of';
-import { merge } from 'rxjs/observable/merge';
-
-/** Todo: Replace this with your own data model type */
-export interface DataItem {
-  name: string;
-  id: number;
-}
-
-function compare(a, b, isAsc) {
-  return (a < b ? -1 : 1) * (isAsc ? 1 : -1);
-}
-
-export class ExampleDataSource extends DataSource<any> {
-  data: DataItem[] = [
-    {id: 1, name: 'Hydrogen'},
-    {id: 2, name: 'Helium'},
-    {id: 3, name: 'Lithium'},
-    {id: 4, name: 'Beryllium'},
-    {id: 5, name: 'Boron'},
-    {id: 6, name: 'Carbon'},
-    {id: 7, name: 'Nitrogen'},
-    {id: 8, name: 'Oxygen'},
-    {id: 9, name: 'Fluorine'},
-    {id: 10, name: 'Neon'},
-    {id: 11, name: 'Sodium'},
-    {id: 12, name: 'Magnesium'},
-    {id: 13, name: 'Aluminum'},
-    {id: 14, name: 'Silicon'},
-    {id: 15, name: 'Phosphorus'},
-    {id: 16, name: 'Sulfur'},
-    {id: 17, name: 'Chlorine'},
-    {id: 18, name: 'Argon'},
-    {id: 19, name: 'Potassium'},
-    {id: 20, name: 'Calcium'},
-  ];
-
-  constructor(private _paginator: MatPaginator, private _sort: MatSort) {
-    super();
-  }
-
-  connect(): Observable<Element[]> {
-    // Combine everything that affects the rendered data into one update
-    // stream for the data-table to consume.
-    const dataMutations = [
-      of(this.data),
-      this._paginator.page,
-      this._sort.sortChange
-    ];
-
-    return merge(...dataMutations).pipe(map(() => {
-      return this.getPagedData(this.getSortedData(this.data));
-    }));
-  }
-
-  disconnect() {}
-
-  private getPagedData(data) {
-    const startIndex = this._paginator.pageIndex * this._paginator.pageSize;
-    return data.splice(startIndex, this._paginator.pageSize);
-  }
-
-  private getSortedData(data) {
-    if (!this._sort.active || this._sort.direction === '') {
-      return data;
-    }
-
-    return data.sort((a, b) => {
-      const isAsc = this._sort.direction == 'asc';
-      switch (this._sort.active) {
-        case 'name': return compare(a.name, b.name, isAsc);
-        case 'id': return compare(+a.id, +b.id, isAsc);
-        default: return 0;
-      }
-    });
-  }
-}
+import { <%= classify(name) %>DataSource } from '<%= dasherize(name) %>-datasource';
 
 @Component({
   selector: '<%= selector %>',<% if(inlineTemplate) { %>
@@ -120,12 +42,12 @@ export class ExampleDataSource extends DataSource<any> {
 export class <%= classify(name) %>Component implements OnInit {
   @ViewChild(MatPaginator) paginator: MatPaginator;
   @ViewChild(MatSort) sort: MatSort;
-  dataSource: ExampleDataSource;
+  dataSource: <%= classify(name) %>DataSource;
 
   /** Columns displayed in the table. Columns IDs can be added, removed, or reordered. */
   displayedColumns = ['id', 'name'];
 
   ngOnInit() {
-    this.dataSource = new ExampleDataSource(this.paginator, this.sort);
+    this.dataSource = new <%= classify(name) %>DataSource(this.paginator, this.sort);
   }
 }
